src/workshop/requirements.txt
httpx>=0.27.2, <0.28.0
aiohttp>=3.11.11, <4.0.0
python_dotenv>=1.0.1, <2.0.0
azure-identity>=1.19.0, <2.0.0
azure-ai-projects>=1.0.0b5
azure-ai-agents>=1.0.0b1
pandas>=2.2.3, <3.0.0
pydantic==2.10.1
pillow>=11.1.0, <12.0.0
pdfminer.six>=20221105  # For PDF text extraction
azure-storage-blob>=12.0.0  # For Azure Blob Storage
aiosqlite>=0.20.0, <1.0.0
flask>=2.2.0
markdown>=3.4.0
bleach>=6.0.0